Remove MergeTree.getStats
