+++ packages/dds/merge-tree/src/test/testClient.ts
@@ -14,7 +14,7 @@ import {
     List,
 } from "../collections";
 import { UnassignedSequenceNumber } from "../constants";
-import { ISegment, Marker } from "../mergeTreeNodes";
+import { IMergeBlock, ISegment, Marker, MaxNodesInBlock, MergeTreeStats } from "../mergeTreeNodes";
 import { createInsertSegmentOp, createRemoveRangeOp } from "../opBuilder";
 import { IJSONSegment, IMarkerDef, IMergeTreeOp, MergeTreeDeltaType, ReferenceType } from "../ops";
 import { PropertySet } from "../properties";
@@ -430,3 +430,49 @@ export const createRevertDriver =
 
     };
 };
+
+export function getStats(tree: MergeTree) {
+    const nodeGetStats = (block: IMergeBlock): MergeTreeStats => {
+        const stats: MergeTreeStats = {
+            maxHeight: 0,
+            nodeCount: 0,
+            leafCount: 0,
+            removedLeafCount: 0,
+            liveCount: 0,
+            histo: [],
+        };
+        for (let k = 0; k < MaxNodesInBlock; k++) {
+            stats.histo[k] = 0;
+        }
+        for (let i = 0; i < block.childCount; i++) {
+            const child = block.children[i];
+            let height = 1;
+            if (!child.isLeaf()) {
+                const childStats = nodeGetStats(child);
+                height = 1 + childStats.maxHeight;
+                stats.nodeCount += childStats.nodeCount;
+                stats.leafCount += childStats.leafCount;
+                stats.removedLeafCount += childStats.removedLeafCount;
+                stats.liveCount += childStats.liveCount;
+                for (let j = 0; j < MaxNodesInBlock; j++) {
+                    stats.histo[j] += childStats.histo[j];
+                }
+            } else {
+                stats.leafCount++;
+                const segment = child;
+                if (segment.removedSeq !== undefined) {
+                    stats.removedLeafCount++;
+                }
+            }
+            if (height > stats.maxHeight) {
+                stats.maxHeight = height;
+            }
+        }
+        stats.histo[block.childCount]++;
+        stats.nodeCount++;
+        stats.liveCount += block.childCount;
+        return stats;
+    };
+    const rootStats = nodeGetStats(tree.root);
+    return rootStats;
+}
